requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-MySQLdb==2.0.0
Flask-pymysql==0.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
Werkzeug==3.1.3
